+++ Flo-News-App/Detail/DetailScreenViewController.swift
@@ -16,11 +16,32 @@ class DetailScreenViewController: UIViewController {
     @IBOutlet weak var contentLbl: UILabel!
     @IBOutlet weak var saveNewsBtn: UIButton!
     @IBOutlet weak var publishDate: UILabel!
+    @IBOutlet weak var author: UILabel!
+    
     var isFav: Bool = false
     var newsResponseModel : Article?
     var dateStringFormat:String?
     
-    func changeFavButton(){
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        setData()
+        getFavNews()
+        formatDate()
+       
+    }
+    
+    private func setData(){
+        image.layer.cornerRadius = 5
+        titleLbl.text = newsResponseModel?.title ?? "News Title"
+        sourceLbl.text = newsResponseModel?.source?.name ?? "BBC"
+        author.text = newsResponseModel?.author ?? "Mehmet Ali Brand"
+        publishDate.text = dateStringFormat
+        contentLbl.text = newsResponseModel?.description ?? "It is a long established fact that a reader will be distracted by the readable content of a page when looking at its layout. The point of using Lorem Ipsum is that it has a more-or-less normal distribution of letters, as opposed to using 'Content here, content here', making it look like readable English"
+    }
+    
+    //MARK: Kaydetme butonuna basınca ikon degismesi
+   private func changeFavButton(){
         if isFav == true {
             
             saveNewsBtn.setImage(UIImage(systemName: "bookmark.fill"), for: .normal)
@@ -28,7 +49,7 @@ class DetailScreenViewController: UIViewController {
             saveNewsBtn.setImage(UIImage(systemName: "bookmark"), for: .normal)
         }
     }
-    @IBOutlet weak var author: UILabel!
+    
     
     @IBAction func backBtn(_ sender: Any) {
         self.dismiss(animated: true)
@@ -49,7 +70,6 @@ class DetailScreenViewController: UIViewController {
             fetchRequest.returnsObjectsAsFaults = false
             do {
                 let results = try context.fetch(fetchRequest)
-                
                 for result in results as! [NSManagedObject] {
                     if let newsID = result.value(forKey: "id")  as? String {
                         print("IDD ==== \(newsID)")
@@ -63,37 +83,37 @@ class DetailScreenViewController: UIViewController {
                 }
                 
             } catch {
-                print("errır")
+                print("error")
             }
         }else{
-            let appDelegate = UIApplication.shared.delegate as! AppDelegate
-            let context = appDelegate.persistentContainer.viewContext
-            
-            let news = NSEntityDescription.insertNewObject(forEntityName: "FavoriteList", into: context)
-            //photo bu coredatadaki isminin adı
-            news.setValue(newsResponseModel?.title ?? "", forKey: "news_title")
-            news.setValue(newsResponseModel?.urlToImage ?? Constants.emptyUrlImage, forKey: "img")
-            news.setValue(newsResponseModel?.publishedAt ?? "", forKey: "id")
-            news.setValue(newsResponseModel?.author ?? "", forKey: "news_author")
-            news.setValue(newsResponseModel?.description ?? "", forKey: "news_content")
-            news.setValue(newsResponseModel?.source?.name ?? "", forKey: "source_name")
-            news.setValue(newsResponseModel?.url ?? "", forKey: "news_url")
-            news.setValue(newsResponseModel?.description ?? "", forKey: "news_description")
-            
-            do {
-                try context.save()
-                isFav = true
-                changeFavButton()
-                print(news)
-                print("SUCCESS")
-            }catch {
-                print("ERROR")
-            }
+          saveFavNews()
         }
     }
-    override func viewDidLoad() {
-        super.viewDidLoad()
-
+    private func saveFavNews(){
+        let appDelegate = UIApplication.shared.delegate as! AppDelegate
+        let context = appDelegate.persistentContainer.viewContext
+        let news = NSEntityDescription.insertNewObject(forEntityName: "FavoriteList", into: context)
+        news.setValue(newsResponseModel?.title ?? "", forKey: "news_title")
+        news.setValue(newsResponseModel?.urlToImage ?? Constants.emptyUrlImage, forKey: "img")
+        news.setValue(newsResponseModel?.publishedAt ?? "", forKey: "id")
+        news.setValue(newsResponseModel?.author ?? "", forKey: "news_author")
+        news.setValue(newsResponseModel?.description ?? "", forKey: "news_content")
+        news.setValue(newsResponseModel?.source?.name ?? "", forKey: "source_name")
+        news.setValue(newsResponseModel?.url ?? "", forKey: "news_url")
+        news.setValue(newsResponseModel?.description ?? "", forKey: "news_description")
+        
+        do {
+            try context.save()
+            isFav = true
+            changeFavButton()
+            print(news)
+            print("SUCCESS")
+        }catch {
+            print("ERROR")
+        }
+    }
+    
+    private func getFavNews(){
         image.kf.setImage(with: URL(string: newsResponseModel?.urlToImage ??  Constants.emptyUrlImage), placeholder:UIImage(named:"image"))
         
         //veri cagırmak eğer haber kaydedilmisse bookmark boyalı olsun
@@ -119,31 +139,28 @@ class DetailScreenViewController: UIViewController {
             }
             
         } catch {
-            print("errır")
+            print("error")
         }
-        
+    }
+    private func formatDate(){
         //gelen tarihi formatlama
         let dateString = newsResponseModel?.publishedAt ?? ""
-
         let dateFormatter = DateFormatter()
         dateFormatter.dateFormat = "yyyy-MM-dd'T'HH:mm:ssZ"
-
+        
         if let date = dateFormatter.date(from: dateString) {
             dateFormatter.dateFormat = "MMMM dd, yyyy"
             dateFormatter.locale = Locale(identifier: "en_US")
             let formattedDate = dateFormatter.string(from: date)
-            print(formattedDate) // May 13, 2020
+            print(formattedDate)
             dateStringFormat = formattedDate
             print("gelen tarihhhhh : \(dateStringFormat)")
         } else {
             print("Invalid date format.")
         }
-        setData()
-        
     }
     
     @IBAction func goToWebView(_ sender: Any) {
-        
         let alertController = UIAlertController(title: "Read More", message: "Do you want to go to the web page?", preferredStyle: .alert)
         
         let yesAction = UIAlertAction(title: "Yes", style: .default) { [self] _ in
@@ -160,14 +177,7 @@ class DetailScreenViewController: UIViewController {
     }
     
     
-    func setData(){
-        image.layer.cornerRadius = 5
-        titleLbl.text = newsResponseModel?.title ?? "News Title"
-        sourceLbl.text = newsResponseModel?.source?.name ?? "BBC"
-        author.text = newsResponseModel?.author ?? "Mehmet Ali Brand"
-        publishDate.text = dateStringFormat
-        contentLbl.text = newsResponseModel?.description ?? "It is a long established fact that a reader will be distracted by the readable content of a page when looking at its layout. The point of using Lorem Ipsum is that it has a more-or-less normal distribution of letters, as opposed to using 'Content here, content here', making it look like readable English"
-    }
+   
     
 }
